[Backend] Adjust the code structure to support the new backend implementation
